feat: Add sourceChain check for airdrop
