+++ src/utils/chains.ts
@@ -1,6 +1,5 @@
-import { IAirdropConfig } from '../exec/types'
-import { ethers } from 'ethers'
-import { ChainInfo } from './types'
+import {ethers} from 'ethers'
+import {ChainInfo, ELayer, IAirdropConfig} from './types'
 
 export type IBobaChain = Omit<ChainInfo, 'chainId' | 'provider'>
 export interface IBobaChains {
@@ -41,6 +40,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_BOBAETHMAINNET ??
       'https://boba-ethereum.gateway.tenderly.co',
     testnet: false,
+    layer: ELayer.Layer2,
     name: 'Boba Ethereum Mainnet',
     teleportationAddress: '0x0dfFd3Efe9c3237Ad7bf94252296272c96237FF5',
     height: 1111267,
@@ -58,6 +58,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_BOBABNBMAINNET ??
       'https://replica.bnb.boba.network',
     testnet: false,
+    layer: ELayer.Layer2,
     name: 'Boba BNB Mainnet',
     teleportationAddress: '0x670b130112C6f03E17192e63c67866e67D77c3ee',
     height: 3393,
@@ -75,6 +76,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_BOBAETHGOERLI ??
       'https://replica.goerli.boba.network',
     testnet: true,
+    layer: ELayer.Layer2,
     name: 'Boba Ethereum Goerli',
     teleportationAddress: '0xB43EE846Aa266228FeABaD1191D6cB2eD9808894',
     height: 40822,
@@ -92,6 +94,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_BOBABNBTESTNET ??
       'https://boba-bnb-testnet.gateway.tenderly.co',
     testnet: true,
+    layer: ELayer.Layer2,
     name: 'Boba BNB Testnet',
     teleportationAddress: '0xf4d179d3a083Fa3Eede935FaF4C679D32d514186',
     height: 295353,
@@ -110,6 +113,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_ARBITRUMGOERLI ??
       'https://arbitrum-goerli.public.blastapi.io',
     testnet: true,
+    layer: ELayer.Layer2,
     name: 'Arbitrum Goerli',
     teleportationAddress: '0x7063f59e1Db3e505D844d11A71C78F92D39E5963',
     height: 53880808,
@@ -126,6 +130,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_OPTIMISMGOERLI ??
       'https://optimism-goerli.publicnode.com',
     testnet: true,
+    layer: ELayer.Layer2,
     name: 'Optimism Goerli',
     teleportationAddress: '0xC226F132A686A08018431C913d87693396246024',
     height: 17010097,
@@ -142,6 +147,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_ARBITRUMMAINNET ??
       'https://arbitrum.llamarpc.com',
     testnet: false,
+    layer: ELayer.Layer2,
     name: 'Arbitrum Mainnet',
     teleportationAddress: '0x2dE73Bd1660Fbf4D521a52Ec2a91CCc106113801',
     height: 180755096,
@@ -158,6 +164,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_OPTIMISMMAINNET ??
       'https://optimism.llamarpc.com',
     testnet: false,
+    layer: ELayer.Layer2,
     name: 'Optimism Mainnet',
     teleportationAddress: '0x2dE73Bd1660Fbf4D521a52Ec2a91CCc106113801',
     height: 116168267,
@@ -175,6 +182,7 @@ export const BobaChains: IBobaChains = {
     url: process.env.LIGHTBRIDGE_RPC_ETHMAINNET ?? 'https://eth.llamarpc.com',
     testnet: false,
     name: 'Ethereum Mainnet',
+    layer: ELayer.Layer1,
     teleportationAddress: '0x2dE73Bd1660Fbf4D521a52Ec2a91CCc106113801',
     height: 19227141,
     supportedAssets: {
@@ -192,6 +200,7 @@ export const BobaChains: IBobaChains = {
     url: process.env.LIGHTBRIDGE_RPC_BNBMAINNET ?? 'https://rpc.ankr.com/bsc',
     testnet: false,
     name: 'BNB Mainnet',
+    layer: ELayer.Layer1,
     teleportationAddress: '0x0dfFd3Efe9c3237Ad7bf94252296272c96237FF5',
     height: 30907682,
     supportedAssets: {
@@ -210,6 +219,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_SEPOLIATESTNET ??
       'https://ethereum-sepolia.publicnode.com',
     testnet: true,
+    layer: ELayer.Layer1,
     name: 'Sepolia Testnet',
     teleportationAddress: '0xaeE12b8D99BBff7ED47866DF868CF5b4b3F73ffF',
     height: 5280795,
@@ -228,6 +238,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_BOBASEPOLIATESTNET ??
       'https://sepolia.boba.network',
     testnet: true,
+    layer: ELayer.Layer2,
     name: 'Boba Sepolia Testnet',
     teleportationAddress: '0x2dE73Bd1660Fbf4D521a52Ec2a91CCc106113801',
     height: 1043907, // TODO: might need to be adapted
@@ -245,6 +256,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_GOERLITESTNET ??
       'https://rpc.ankr.com/eth_goerli',
     testnet: true,
+    layer: ELayer.Layer1,
     name: 'Goerli Testnet',
     teleportationAddress: '0x84b22166366a6f7E0cD0c3ce9998f2913Bf17A13',
     height: 9484025,
@@ -264,6 +276,7 @@ export const BobaChains: IBobaChains = {
       process.env.LIGHTBRIDGE_RPC_BNBTESTNET ??
       'https://api.zan.top/node/v1/bsc/testnet/public',
     testnet: true,
+    layer: ELayer.Layer1,
     name: 'BNB Testnet',
     teleportationAddress: '0x7f6a32bCaA70c65E08F2f221737612F6fC18347A',
     height: 32272487,
